requirements.txt
akshare>=1.13.0
pandas>=2.0.0
requests>=2.31.0
ta>=0.11.0
pyyaml>=6.0
lxml>=4.9.0
scipy>=1.10.0